updated 5x5 computer player to skip algoritm on first move
Recursive depth is not sufficient to find a win or loss on
the first move, so calling the algoritm is a waste of time.
Instead, if the middle box is available, it goes there. If
the middle is taken, if any of the 4 corners are available,
it picks one at random. Otherwise, it picks a random spot
on the board.